requirements.txt
tensorflow-cpu==2.15.1
tokenizers==0.14
fastapi
uvicorn
python-multipart
numpy
pillow
requests
google-cloud-dialogflow
google-auth